requirements.txt
atlassian-python-api==3.32.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
et-xmlfile==1.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openpyxl==3.1.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.14
wrapt==1.14.1
xlrd==2.0.1
xlsxtpl==0.3.1
xltpl==0.17
xlwt==1.3.0
